brain_tumour_classification/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.4
Pillow==12.1.1
scikit-learn==1.7.2
mlflow==2.9.2
azureml-mlflow==1.54.0
azure-ai-ml==1.32.0
azureml-fsspec==1.3.1
azureml-core==1.61.0.post3
azureml-dataset-runtime==1.62.0
